bug fix
Former-commit-id: 70f1d2f18a1e0eb1cb071595f581e0d99793a436
Former-commit-id: 276e94dc27a26fbbd60ad9cf5638411e896f414b
Former-commit-id: 4ba6c0f242b47bacf37ece7d73d8e464cc994f72
Former-commit-id: cdd12de04b27527e33dababa2a601eebc6613372 [formerly f1f87a603e8b653b24eb3cb510ff77fb981c9fdc]
Former-commit-id: 4c148cc118204cb20ba801310bbf7d79f6e413b9

+++ src/tamm/types.hpp
@@ -67,7 +67,7 @@ using Size       = Offset;
 using BlockCount = BlockIndex;
 using Label      = int; //needs to support negative values
 
-using IntLabel = size_t; //a simple integer label for indices
+using IntLabel = int32_t; //a simple integer label for indices
 using IntLabelVec = std::vector<IntLabel>;
 
 using SizeVec = std::vector<Size>;
